requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
websockets==12.0
ccxt==4.2.48
tensorflow==2.16.1
numpy==1.24.3
pandas==2.1.4
scikit-learn==1.3.2
xgboost==2.0.2
redis==5.0.1
bcrypt==4.1.2
pyjwt==2.8.0
python-dotenv==1.0.0
speakeasy==2.0.0
qrcode==7.4.2
technicalindicators==3.1.0
